rename filereader.xx->logreader.xx
